requirements.txt
python-telegram-bot[webhooks]==20.8
requests
python-dotenv
aiohttp
colorlog